+++ CHANGES.md
@@ -1,4 +1,5 @@
 # 6.5.1 2023-??-??
+ - 2022-12-16 Added new function Kernel::System::Web::Request::GetParams() to get request parameters.
  - 2022-12-13 Added non-standard time zone mapping for iCal/ICS handling.
  - 2022-12-12 Fixed bulk un-mention action via mention view link.
  - 2022-12-07 Added independent ToolBar::TicketProcess. This is only visible when a valid process exists.